+++ cmd/api/request/add_to_cart_request.go
@@ -1,6 +1,6 @@
 package request
 
 type AddToCartRequest struct {
-	ProductID int64 `json:"product_id" validate:"required"`
+	ProductID int `json:"product_id" validate:"required"`
 	Quantity  int   `json:"quantity" validate:"required,gt=0"`
 }
