+++ cmd/api/request/update_cart_request.go
@@ -1,5 +1,5 @@
 package request
 
 type UpdateCartRequest struct {
-	Quantity int64 `json:"quantity" validate:"required"`
+	Quantity int `json:"quantity" validate:"required"`
 }
